requirements.txt
python-dotenv==0.14.0
spotipy==2.13.0
Flask==1.1.1